Common: Increase parameters to handle millions of sprites,

+++ Source/Fodder.cpp
@@ -2544,7 +2544,7 @@ void cFodder::Sprite_Sort_DrawList() {
     for(sSprite& sprite : mSprites) {
 
         if (sprite.field_0 == -32768)
-            continue;
+            continue; 
 
         if (sprite.field_2C == eSprite_Draw_First) {
             mSprite_DrawList_First.push_back(&sprite);
@@ -10179,7 +10179,7 @@ void cFodder::Sprite_Frame_Modifier_Update() {
 void cFodder::Sprite_Handle_Loop() {
     sSprite* Data20 = mSprites.data();
 
-    for (int16 Data1C = mParams.mSpritesMax - 2; Data1C > 0; --Data1C, ++Data20) {
+    for (int32 Data1C = mParams.mSpritesMax - 2; Data1C > 0; --Data1C, ++Data20) {
 
         if (Data20->field_0 == -32768)
             continue;
@@ -17236,7 +17236,7 @@ int16 cFodder::Sprite_Get_Free_Max42(int16& pData0, sSprite*& pData2C, sSprite*&
             pData2C = mSprites.data();
 
             // Loop all sprites
-            for (int16 Data1C = mParams.mSpritesMax - 5; Data1C >= 0; --Data1C, ++pData2C) {
+             for (int32_t Data1C = mParams.mSpritesMax - 5; Data1C >= 0; --Data1C, ++pData2C) {
 
                 // Sprite free?
                 if (pData2C->field_0 != -32768)
@@ -17260,7 +17260,7 @@ int16 cFodder::Sprite_Get_Free_Max42(int16& pData0, sSprite*& pData2C, sSprite*&
             pData2C = mSprites.data();
 
             // Loop all sprites
-            for (int16 Data1C = mParams.mSpritesMax - 4; Data1C >= 0; --Data1C, ++pData2C) {
+             for (int32_t Data1C = mParams.mSpritesMax - 4; Data1C >= 0; --Data1C, ++pData2C) {
 
                 // Sprite free?
                 if (pData2C->field_0 != -32768)
@@ -17281,7 +17281,7 @@ int16 cFodder::Sprite_Get_Free_Max42(int16& pData0, sSprite*& pData2C, sSprite*&
             // Only looking for 1 sprite
             pData2C = &mSprites[mParams.mSpritesMax - 3];
 
-            for (int16 Data1C = mParams.mSpritesMax - 3; Data1C >= 0; --Data1C) {
+            for (int32 Data1C = mParams.mSpritesMax - 3; Data1C >= 0; --Data1C) {
 
                 // Free?
                 if (pData2C->field_0 == -32768) {
@@ -17311,7 +17311,7 @@ int16 cFodder::Sprite_Get_Free_Max29(int16& pData0, sSprite*& pData2C, sSprite*&
         goto loc_21B91;
 
     pData2C = &mSprites[mParams.mSpritesMax - 16];
-    for (int16 Data1C = mParams.mSpritesMax - 16; Data1C >= 0; --Data1C, --pData2C) {
+     for (int32_t Data1C = mParams.mSpritesMax - 16; Data1C >= 0; --Data1C, --pData2C) {
 
         if (pData2C->field_0 == -32768) {
             Sprite_Clear(pData2C);
@@ -17330,7 +17330,7 @@ loc_21B4B:;
 loc_21B91:;
     pData2C = mSprites.data();
 
-    for (int16 Data1C = mParams.mSpritesMax - 17; Data1C >= 0; --Data1C, ++pData2C) {
+     for (int32_t Data1C = mParams.mSpritesMax - 17; Data1C >= 0; --Data1C, ++pData2C) {
 
         if (pData2C->field_0 != -32768)
             continue;
